Move TestIsPackagingWorkflow to fileparser. Add minimal test for pypa/gh-action-pypi-publish.
